refactor a lot

+++ app/controllers/users_controller.rb
@@ -18,18 +18,17 @@ class UsersController < ApplicationController
   end
 
   def toggle
-    puts "current user starts at #{current_user.moments_on}"
     current_user.moments_on = !current_user.moments_on
-    puts "current user is receiving moments: #{current_user.moments_on}"
-    current_user.save
-    if (current_user.moments_on && UserQuery.new(current_user).missing_moments?)
-      puts "making moments for user from toggle"
-      Moment.generate_moments_for_day(current_user)
-    else  
-      puts "destroying all unsent momvents"
-      MomentsNotSentQuery.new(current_user).search.destroy_all
-    end 
-    render json: {moments_on: current_user.moments_on}
+    if current_user.save
+      if (current_user.moments_on && UserQuery.new(current_user).missing_moments?)
+        Moment.generate_moments_for_day(current_user)
+      else  
+        MomentsNotSentQuery.new(current_user).search.destroy_all
+      end 
+      render json: {moments_on: current_user.moments_on}
+     else
+      render json: {moments_on: false, errors: current_user.errors.messages}
+    end
   end
 
   def time
@@ -38,11 +37,10 @@ class UsersController < ApplicationController
     current_user.start_time = params['start_time']
     current_user.end_time = params['end_time']
       if current_user.save
-        puts "in the save ajax"
-        if UserQuery.new(current_user).no_moments_today?
+        if current_user.no_moments_today?
           Moment.generate_moments_for_day(current_user)
         end 
-        render json: {success: true, start_time: current_user.start_time, end_time: current_user.end_time}.to_json
+        render json: {success: true, start_time: current_user.start_time, end_time: current_user.end_time}
       else 
         render json: {success: false, errors: current_user.errors.messages, start_time: current_user.start_time, end_time: current_user.end_time }
       end
